sample-client/requirements.txt
faker
requests
python-dotenv